change `mod env` location to `session::` & send close frame when timeout
